# libs/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@daytona.com
    name: Daytona Platforms Inc.
    url: https://www.daytona.io
  description: Daytona AI platform API Docs
  title: Daytona
  version: '1.0'
servers:
  - url: http://localhost:3000
paths:
...
  /sandbox:
    get:
      operationId: listSandboxes
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Include verbose output
          explode: true
          in: query
          name: verbose
          required: false
          schema:
            type: boolean
          style: form
        - description: JSON encoded labels to filter by
          explode: true
          in: query
          name: labels
          required: false
          schema:
            example: '{"label1": "value1", "label2": "value2"}'
            type: string
          style: form
        - description: Include errored and deleted sandboxes
          explode: true
          in: query
          name: includeErroredDeleted
          required: false
          schema:
            type: boolean
          style: form
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Sandbox'
                type: array
          description: List of all sandboxes
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: List all sandboxes
      tags:
        - sandbox
    post:
      operationId: createSandbox
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSandbox'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
          description: The sandbox has been successfully created.
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Create a new sandbox
      tags:
        - sandbox
  /sandbox/{sandboxId}:
    delete:
      operationId: deleteSandbox
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
        - explode: true
          in: query
          name: force
          required: true
          schema:
            type: boolean
          style: form
      responses:
        '200':
          description: Sandbox has been deleted
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Delete sandbox
      tags:
        - sandbox
    get:
      operationId: getSandbox
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Include verbose output
          explode: true
          in: query
          name: verbose
          required: false
          schema:
            type: boolean
          style: form
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
          description: Sandbox details
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Get sandbox details
      tags:
        - sandbox
  /sandbox/{sandboxId}/start:
    post:
      operationId: startSandbox
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
          description: Sandbox has been started
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Start sandbox
      tags:
        - sandbox
  /sandbox/{sandboxId}/stop:
    post:
      operationId: stopSandbox
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          description: Sandbox has been stopped
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Stop sandbox
      tags:
        - sandbox
  /sandbox/{sandboxId}/labels:
    put:
      operationId: replaceLabels
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SandboxLabels'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxLabels'
          description: Labels have been successfully replaced
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Replace sandbox labels
      tags:
        - sandbox
  /sandbox/{sandboxId}/backup:
    post:
      operationId: createBackup
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
          description: Sandbox backup has been initiated
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Create sandbox backup
      tags:
        - sandbox
  /sandbox/{sandboxId}/public/{isPublic}:
    post:
      operationId: updatePublicStatus
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
        - description: Public status to set
          explode: false
          in: path
          name: isPublic
          required: true
          schema:
            type: boolean
          style: simple
      responses:
        '201':
          description: ''
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Update public status
      tags:
        - sandbox
  /sandbox/{sandboxId}/autostop/{interval}:
    post:
      operationId: setAutostopInterval
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
        - description: Auto-stop interval in minutes (0 to disable)
          explode: false
          in: path
          name: interval
          required: true
          schema:
            type: number
          style: simple
      responses:
        '200':
          description: Auto-stop interval has been set
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Set sandbox auto-stop interval
      tags:
        - sandbox
  /sandbox/{sandboxId}/autoarchive/{interval}:
    post:
      operationId: setAutoArchiveInterval
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the sandbox
          explode: false
          in: path
          name: sandboxId
          required: true
          schema:
            type: string
          style: simple
        - description: Auto-archive interval in minutes (0 means the maximum interval
            will be used)
          explode: false
          in: path
          name: interval
          required: true
          schema:
            type: number
          style: simple
      responses:
        '200':
          description: Auto-archive interval has been set
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Set sandbox auto-archive interval
      tags:
        - sandbox
...
components:
  schemas:
...
    SandboxLabels:
      example:
        labels:
          environment: dev
          team: backend
      properties:
        labels:
          additionalProperties:
            type: string
          description: Key-value pairs of labels
          example:
            environment: dev
            team: backend
          type: object
      required:
        - labels
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect